Develop

+++ src/main/resources/static/js/userService.js
@@ -16,7 +16,7 @@ async function getUsers() {
         return `
         <tr>
           <td>${user.id}</td>
-          <td>${user.name + " " + user.lastName}</td>
+          <td>${user.firstName + " " + user.lastName}</td>
           <td>${user.email}</td>
           <td>${user.phone}</td>
           <td>
@@ -37,18 +37,24 @@ async function getUsers() {
   }
 }
 
+
+
 async function deleteUser(id) {
   try {
-    const request = await fetch(`http://localhost:8080/api/users/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-    if (!request.ok) {
-      throw new Error(`HTTP error! status: ${request.status}`);
+    const delete_confirmation = confirm("Are you sure you want to delete?");
+    if (delete_confirmation) {
+      const request = await fetch(`http://localhost:8080/api/users/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+
+      if (!request.ok) {
+        throw new Error(`HTTP error! status: ${request.status}`);
+      }
+      getUsers();
     }
-    getUsers();
   } catch (error) {
     console.error("Error deleting user:", error);
   }
